Introduce virtual loss mechanism
A virtual loss is simply an extra visit with a zero score. It is added
by the tree search algorithm to each node visited by a thread during the
selection step, and removed again after simulation. This mechanism makes
it temporarily less likely for other threads to select the same sequence
of nodes.