Fix determinisation algorithm issues
It was not guaranteed to succeed, depending on the number of active hand
constraints. This led the simulated games to fail early and report bogus
scores. Dealing is not deterministic, though semi-intelligent and fast
and it should only rarely take more than a dozen tries. The arbitrary
cutoff at 1000 attempts should therefore be sufficient.